test: add basic tests on the data displayed in the component
