+++ springwolf-ui/src/app/components/channels/channel-main/channel-main.component.spec.ts
@@ -4,28 +4,16 @@ import { ChannelMainComponent } from "./channel-main.component";
 import { AsyncApiService } from "../../../service/asyncapi/asyncapi.service";
 import { PublisherService } from "../../../service/publisher.service";
 import { MaterialModule } from "../../../material.module";
-import { of } from "rxjs";
-import { exampleSchemas } from "../../../service/mock/example-data";
-import { AsyncApiMapperService } from "../../../service/asyncapi/asyncapi-mapper.service";
 import { MarkdownModule } from "ngx-markdown";
 import {
   mockedAsyncApiService,
   mockedExampleSchemaMapped,
 } from "../../../service/mock/mock-asyncapi.service";
-import { Component, Input } from "@angular/core";
-import { Schema } from "../../../models/schema.model";
-
-@Component({ selector: "app-json", template: "" })
-export class MockAppJson {
-  @Input() data: any;
-}
-
-@Component({ selector: "app-schema", template: "" })
-export class MockAppSchema {
-  @Input() schema: Schema;
-}
+import { MockAppJson, MockAppSchema } from "../../mock-components.spec";
 
 describe("ChannelMainComponent", () => {
+  const mockData = mockedExampleSchemaMapped.channelOperations[0];
+
   beforeEach(async () => {
     mockedAsyncApiService.getAsyncApi.mockClear();
 
@@ -37,13 +25,15 @@ describe("ChannelMainComponent", () => {
         { provide: PublisherService, useValue: {} },
       ],
       componentProperties: {
-        channelName: mockedExampleSchemaMapped.channelOperations[0].name,
-        operation: mockedExampleSchemaMapped.channelOperations[0].operation,
+        channelName: mockData.name,
+        operation: mockData.operation,
       },
     });
   });
 
-  it("should create the component", () => {
-    expect(screen.getByText("Another payload model")).toBeTruthy();
+  it("should render the component and data", () => {
+    expect(
+      screen.getByText(mockData.operation.message.description)
+    ).toBeTruthy();
   });
 });
